+++ scl/cxx/inferenceModule/test/units/TestManagerModule.cpp
@@ -16,10 +16,11 @@ using namespace inference;
 namespace directInferenceManagerTest
 {
 ScsLoader loader;
-const std::string TEST_FILES_DIR_PATH = TEMPLATE_SEARCH_MODULE_TEST_SRC_PATH "/testStructures/ManagerModule/";
-const std::string TEMPLATE_REPLY_TARGET = "template_reply_target";
-const std::string RULE_CLASS = "concept_answer_on_standard_message_rule_class_by_priority";
-const std::string MESSAGE_SINGLETON = "message_singleton";
+std::string const TEST_FILES_DIR_PATH = TEMPLATE_SEARCH_MODULE_TEST_SRC_PATH "/testStructures/ManagerModule/";
+
+std::string const TARGET_TEMPLATE = "target_template";
+std::string const RULES_SET = "rules_set";
+std::string const ARGUMENT_SET = "argument_set";
 
 using InferenceManagerTest = ScMemoryTest;
 
@@ -29,189 +30,77 @@ void initialize()
   scAgentsCommon::CoreKeynodes::InitGlobal();
 }
 
-// TODO (MksmOrlov): fail old interface
 TEST_F(InferenceManagerTest, SuccessApplyInference)
 {
   ScMemoryContext & context = *m_ctx;
 
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "lr_concept_weather_request_message_1.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "inferenceManagerRuleUsedTest.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "concept_answer_on_standard_message_rule_class_by_priority.scs");
+  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "trueSimpleRuleTest.scs");
   initialize();
 
-  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TEMPLATE_REPLY_TARGET);
+  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TARGET_TEMPLATE);
   EXPECT_TRUE(targetTemplate.IsValid());
 
-  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULE_CLASS);
+  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULES_SET);
   EXPECT_TRUE(ruleSet.IsValid());
 
-  ScAddr argumentSet = context.HelperResolveSystemIdtf(MESSAGE_SINGLETON);
+  ScAddr argumentSet = context.HelperResolveSystemIdtf(ARGUMENT_SET);
   EXPECT_TRUE(argumentSet.IsValid());
 
+  ScAddr outputStructure = context.CreateNode(ScType::NodeConstStruct);
   DirectInferenceManager inferenceManager(&context);
-  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet);
+  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet, outputStructure);
   EXPECT_TRUE(answer.IsValid());
   EXPECT_TRUE(
       context.HelperCheckEdge(InferenceKeynodes::concept_success_solution, answer, ScType::EdgeAccessConstPosPerm));
 }
 
-TEST_F(InferenceManagerTest, RuleNotFound)
-{
-  ScMemoryContext & context = *m_ctx;
-
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "inferenceManagerRuleUsedTest.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "incorrect_rule_class_logic_rule.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "concept_answer_on_standard_message_rule_class_by_priority.scs");
-  initialize();
-
-  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TEMPLATE_REPLY_TARGET);
-  EXPECT_TRUE(targetTemplate.IsValid());
-
-  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULE_CLASS);
-  EXPECT_TRUE(ruleSet.IsValid());
-
-  ScAddr argumentSet = context.HelperResolveSystemIdtf(MESSAGE_SINGLETON);
-  EXPECT_TRUE(argumentSet.IsValid());
-
-  DirectInferenceManager inferenceManager(&context);
-  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet);
-
-  EXPECT_TRUE(answer.IsValid());
-  EXPECT_TRUE(
-      context.HelperCheckEdge(InferenceKeynodes::concept_success_solution, answer, ScType::EdgeAccessConstNegPerm));
-}
-
-// TODO (MksmOrlov): fail old interface
 TEST_F(InferenceManagerTest, RuleNotUsed)
 {
   ScMemoryContext & context = *m_ctx;
 
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "inferenceManagerRuleNotUsedTest.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "lr_concept_weather_request_message_1.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "concept_answer_on_standard_message_rule_class_by_priority.scs");
+  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "ruleNotUsedTest.scs");
   initialize();
 
-  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TEMPLATE_REPLY_TARGET);
+  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TARGET_TEMPLATE);
   EXPECT_TRUE(targetTemplate.IsValid());
 
-  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULE_CLASS);
+  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULES_SET);
   EXPECT_TRUE(ruleSet.IsValid());
 
-  ScAddr argumentSet = context.HelperResolveSystemIdtf(MESSAGE_SINGLETON);
+  ScAddr argumentSet = context.HelperResolveSystemIdtf(ARGUMENT_SET);
   EXPECT_TRUE(argumentSet.IsValid());
 
+  ScAddr outputStructure = context.CreateNode(ScType::NodeConstStruct);
   DirectInferenceManager inferenceManager(&context);
-  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet);
-
+  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet, outputStructure);
   EXPECT_TRUE(answer.IsValid());
   EXPECT_TRUE(
       context.HelperCheckEdge(InferenceKeynodes::concept_success_solution, answer, ScType::EdgeAccessConstNegPerm));
 }
 
-TEST_F(InferenceManagerTest, IncorrectElseStatementStructure)
+TEST_F(InferenceManagerTest, TargetNotAchieved)
 {
   ScMemoryContext & context = *m_ctx;
 
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "inferenceManagerRuleUsedTest.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "incorrect_else_statement_logic_rule.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "concept_answer_on_standard_message_rule_class_by_priority.scs");
+  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "targetNotAchievedTest.scs");
   initialize();
 
-  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TEMPLATE_REPLY_TARGET);
+  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TARGET_TEMPLATE);
   EXPECT_TRUE(targetTemplate.IsValid());
 
-  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULE_CLASS);
+  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULES_SET);
   EXPECT_TRUE(ruleSet.IsValid());
 
-  ScAddr argumentSet = context.HelperResolveSystemIdtf(MESSAGE_SINGLETON);
+  ScAddr argumentSet = context.HelperResolveSystemIdtf(ARGUMENT_SET);
   EXPECT_TRUE(argumentSet.IsValid());
 
+  ScAddr outputStructure = context.CreateNode(ScType::NodeConstStruct);
   DirectInferenceManager inferenceManager(&context);
-  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet);
+  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet, outputStructure);
 
   EXPECT_TRUE(answer.IsValid());
   EXPECT_TRUE(
       context.HelperCheckEdge(InferenceKeynodes::concept_success_solution, answer, ScType::EdgeAccessConstNegPerm));
 }
 
-// TODO (MksmOrlov): infinite cycle
-TEST_F(InferenceManagerTest, IncorrectIfStatementStructure)
-{
-  ScMemoryContext & context = *m_ctx;
-
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "inferenceManagerRuleUsedTest.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "incorrect_if_statement_logic_rule.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "concept_answer_on_standard_message_rule_class_by_priority.scs");
-  initialize();
-
-  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TEMPLATE_REPLY_TARGET);
-  EXPECT_TRUE(targetTemplate.IsValid());
-
-  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULE_CLASS);
-  EXPECT_TRUE(ruleSet.IsValid());
-
-  ScAddr argumentSet = context.HelperResolveSystemIdtf(MESSAGE_SINGLETON);
-  EXPECT_TRUE(argumentSet.IsValid());
-
-  DirectInferenceManager inferenceManager(&context);
-  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet);
-
-  EXPECT_TRUE(answer.IsValid());
-  EXPECT_TRUE(
-      context.HelperCheckEdge(InferenceKeynodes::concept_success_solution, answer, ScType::EdgeAccessConstNegPerm));
-}
-
-// TODO (MksmOrlov): fail old interface
-TEST_F(InferenceManagerTest, IncorrectTargetTemplate)
-{
-  ScMemoryContext & context = *m_ctx;
-
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "inferenceManagerIncorrectTemplateTest.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "lr_concept_weather_request_message_1.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "concept_answer_on_standard_message_rule_class_by_priority.scs");
-  initialize();
-
-  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TEMPLATE_REPLY_TARGET);
-  EXPECT_TRUE(targetTemplate.IsValid());
-
-  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULE_CLASS);
-  EXPECT_TRUE(ruleSet.IsValid());
-
-  ScAddr argumentSet = context.HelperResolveSystemIdtf(MESSAGE_SINGLETON);
-  EXPECT_TRUE(argumentSet.IsValid());
-
-  DirectInferenceManager inferenceManager(&context);
-  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet);
-
-  EXPECT_TRUE(answer.IsValid());
-  EXPECT_TRUE(
-      context.HelperCheckEdge(InferenceKeynodes::concept_success_solution, answer, ScType::EdgeAccessConstNegPerm));
-}
-
-// TODO (MksmOrlov): fail old interface
-TEST_F(InferenceManagerTest, IncorrectArgumentSet)
-{
-  ScMemoryContext & context = *m_ctx;
-
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "inferenceManagerIncorrectArgumentSetTest.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "lr_concept_weather_request_message_1.scs");
-  loader.loadScsFile(context, TEST_FILES_DIR_PATH + "concept_answer_on_standard_message_rule_class_by_priority.scs");
-  initialize();
-
-  ScAddr targetTemplate = context.HelperResolveSystemIdtf(TEMPLATE_REPLY_TARGET);
-  EXPECT_TRUE(targetTemplate.IsValid());
-
-  ScAddr ruleSet = context.HelperResolveSystemIdtf(RULE_CLASS);
-  EXPECT_TRUE(ruleSet.IsValid());
-
-  ScAddr argumentSet = context.HelperResolveSystemIdtf(MESSAGE_SINGLETON);
-  EXPECT_TRUE(argumentSet.IsValid());
-
-  DirectInferenceManager inferenceManager(&context);
-  ScAddr answer = inferenceManager.applyInference(targetTemplate, ruleSet, argumentSet);
-
-  EXPECT_TRUE(answer.IsValid());
-  EXPECT_TRUE(
-      context.HelperCheckEdge(InferenceKeynodes::concept_success_solution, answer, ScType::EdgeAccessConstNegPerm));
-}
 }  // namespace directInferenceManagerTest